Move CDMValidator
